backend/data-management/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-dotenv==1.0.0
aiohttp==3.9.0
pydantic==2.5.0
gunicorn==21.2.0
boto3==1.34.0
rq==1.15.1